requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
streamlit>=1.25.0
plotly>=5.14.0
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
umap-learn>=0.5.3
sentencepiece>=0.1.99

